+++ pcs/test/tools/command_env/config_runner_cib.py
@@ -25,6 +25,7 @@ class CibShortcuts:
         before=None,
         returncode=0,
         stderr=None,
+        instead=None,
         **modifier_shortcuts
     ):
         """
@@ -35,6 +36,10 @@ class CibShortcuts:
             returns new etree.Element with desired modification.
         string filename -- points to file with cib in the content
         string before -- key of call before which this new call is to be placed
+        int returncode
+        string stderr
+        string instead -- key of call instead of which this new call is to be
+            placed
         dict modifier_shortcuts -- a new modifier is generated from each
             modifier shortcut.
             As key there can be keys of MODIFIER_GENERATORS.
@@ -66,7 +71,7 @@ class CibShortcuts:
             )
             call = RunnerCall(command, stdout=cib)
 
-        self.__calls.place(name, call, before=before)
+        self.__calls.place(name, call, before=before, instead=instead)
 
     def load_content(
         self,