+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - Command `pcs config checkpoint diff` for displaying differences between two
   specified checkpoints ([rhbz#1655055])
+- Support for resource instance attributes uniqueness check according to
+  resource agent metadata ([rhbz#1665404])
 
 ### Fixed
 - Corosync config file parser updated and made more strict to match changes in
@@ -23,6 +25,7 @@
 [rhbz#1656953]: https://bugzilla.redhat.com/show_bug.cgi?id=1656953
 [rhbz#1659051]: https://bugzilla.redhat.com/show_bug.cgi?id=1659051
 [rhbz#1664057]: https://bugzilla.redhat.com/show_bug.cgi?id=1664057
+[rhbz#1665404]: https://bugzilla.redhat.com/show_bug.cgi?id=1665404
 
 
 ## [0.10.1] - 2018-11-23